feat: implements transactionalId

+++ product-api/src/main/java/br/com/microservico/product_api/modules/product/service/ProductService.java
@@ -8,6 +8,7 @@ import br.com.microservico.product_api.modules.product.model.Product;
 import br.com.microservico.product_api.modules.product.repository.ProductRepository;
 import br.com.microservico.product_api.modules.sales.client.SalesClient;
 import br.com.microservico.product_api.modules.sales.dto.SalesConfirmationDTO;
+import br.com.microservico.product_api.modules.sales.dto.SalesProductResponse;
 import br.com.microservico.product_api.modules.sales.enums.SalesStatusEnum;
 import br.com.microservico.product_api.modules.sales.rabbit.SalesConfirmationSender;
 import br.com.microservico.product_api.modules.supplier.service.SupplierService;
@@ -16,17 +17,24 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import static br.com.microservico.product_api.config.exception.interceptor.RequestUtil.getCurrentRequest;
 import static org.springframework.util.ObjectUtils.isEmpty;
 
 @Slf4j
 @Service
 public class ProductService {
+
+    private static final Integer ZERO = 0;
+    private static final String AUTHORIZATION = "Authorization";
+    private static final String TRANSACTION_ID = "transactionid";
+    private static final String SERVICE_ID = "serviceid";
 
     @Autowired
     private ProductRepository productRepository;
@@ -38,41 +46,91 @@ public class ProductService {
     private SalesConfirmationSender salesConfirmationSender;
     @Autowired
     private SalesClient salesClient;
+    private final ObjectMapper objectMapper;
 
-    public ProductResponse save(ProductRequest request) {
-        validadeProductObject(request);
-        var supplier = supplierService.verifySupplierExists(request.getSupplier());
-        var category = categoryService.verifyCategoryExists(request.getCategory());
-        var product = productRepository.save(Product.of(request, supplier, category));
-        return ProductResponse.of(productRepository.save(product));
+    public ProductService(ObjectMapper objectMapper) {
+        this.objectMapper = objectMapper;
     }
 
-    public ProductResponse update(ProductRequest request, Integer id) {
-        validadeProductObject(request);
-        validateInformedId(id);
-
-        var supplier = supplierService.verifySupplierExists(request.getSupplier());
-        var category = categoryService.verifyCategoryExists(request.getCategory());
-        var product = Product.of(request, supplier, category);
-        product.setId(id);
-
-        productRepository.save(product);
-
+    public ProductResponse save(ProductRequest request) {
+        validateProductDataInformed(request);
+        validateCategoryAndSupplierIdInformed(request);
+        var category = categoryService.findById(request.getCategory());
+        var supplier = supplierService.findById(request.getSupplier());
+        var product = productRepository.save(Product.of(request, supplier, category));
         return ProductResponse.of(product);
     }
 
-    public List<ProductResponse> findAllProducts() {
-        return productRepository.findAll()
+    public ProductResponse update(ProductRequest request,
+                                  Integer id) {
+        validateProductDataInformed(request);
+        validateInformedId(id);
+        validateCategoryAndSupplierIdInformed(request);
+        var category = categoryService.findById(request.getCategory());
+        var supplier = supplierService.findById(request.getSupplier());
+        var product = Product.of(request, supplier, category);
+        product.setId(id);
+        productRepository.save(product);
+        return ProductResponse.of(product);
+    }
+
+    private void validateProductDataInformed(ProductRequest request) {
+        if (isEmpty(request.getName())) {
+            throw new ValidationException("The product's name was not informed.");
+        }
+        if (isEmpty(request.getQuantity())) {
+            throw new ValidationException("The product's quantity was not informed.");
+        }
+        if (request.getQuantity() <= ZERO) {
+            throw new ValidationException("The quantity should not be less or equal to zero.");
+        }
+    }
+
+    private void validateCategoryAndSupplierIdInformed(ProductRequest request) {
+        if (isEmpty(request.getCategory())) {
+            throw new ValidationException("The category ID was not informed.");
+        }
+        if (isEmpty(request.getSupplier())) {
+            throw new ValidationException("The supplier ID was not informed.");
+        }
+    }
+
+    public List<ProductResponse> findAll() {
+        return productRepository
+                .findAll()
                 .stream()
                 .map(ProductResponse::of)
-                .toList();
+                .collect(Collectors.toList());
     }
 
     public List<ProductResponse> findByName(String name) {
         if (isEmpty(name)) {
-            throw new ValidationException("Product Name must be informed");
+            throw new ValidationException("The product name must be informed.");
         }
-        return productRepository.findByName(name)
+        return productRepository
+                .findByName(name)
+                .stream()
+                .map(ProductResponse::of)
+                .collect(Collectors.toList());
+    }
+
+    public List<ProductResponse> findBySupplierId(Integer supplierId) {
+        if (isEmpty(supplierId)) {
+            throw new ValidationException("The product' supplier ID name must be informed.");
+        }
+        return productRepository
+                .findBySupplierId(supplierId)
+                .stream()
+                .map(ProductResponse::of)
+                .collect(Collectors.toList());
+    }
+
+    public List<ProductResponse> findByCategoryId(Integer categoryId) {
+        if (isEmpty(categoryId)) {
+            throw new ValidationException("The product' category ID name must be informed.");
+        }
+        return productRepository
+                .findByCategoryId(categoryId)
                 .stream()
                 .map(ProductResponse::of)
                 .collect(Collectors.toList());
@@ -82,7 +140,6 @@ public class ProductService {
         return ProductResponse.of(findById(id));
     }
 
-
     public Product findById(Integer id) {
         validateInformedId(id);
         return productRepository
@@ -90,54 +147,60 @@ public class ProductService {
                 .orElseThrow(() -> new ValidationException("There's no product for the given ID."));
     }
 
-    public List<ProductResponse> findBySupplierId(Integer supplierId) {
-        validateInformedId(supplierId);
-
-        return productRepository
-                .findBySupplierId(supplierId)
-                .stream()
-                .map(ProductResponse::of)
-                .collect(Collectors.toList());
+    public Boolean existsByCategoryId(Integer categoryId) {
+        return productRepository.existsByCategoryId(categoryId);
     }
 
-    public List<ProductResponse> findByCategoryId(Integer categoryId) {
-        validateInformedId(categoryId);
-        return productRepository
-                .findByCategoryId(categoryId)
-                .stream()
-                .map(ProductResponse::of)
-                .collect(Collectors.toList());
-    }
-
-    public Boolean existsByCategoryId(Integer id) {
-        return productRepository.existsByCategoryId(id);
-    }
-
-    public Boolean existsBySupplierId(Integer id) {
-        return productRepository.existsBySupplierId(id);
+    public Boolean existsBySupplierId(Integer supplierId) {
+        return productRepository.existsBySupplierId(supplierId);
     }
 
     public SuccesResoponse delete(Integer id) {
         validateInformedId(id);
+        if (!productRepository.existsById(id)) {
+            throw new ValidationException("The product does not exists.");
+        }
+        var sales = getSalesByProductId(id);
+        if (!isEmpty(sales.getSalesId())) {
+            throw new ValidationException("The product cannot be deleted. There are sales for it.");
+        }
         productRepository.deleteById(id);
-        return SuccesResoponse.create("The product has been deleted");
+        return SuccesResoponse.create("The product was deleted.");
     }
 
-    public void validateInformedId(Integer id) {
+    private void validateInformedId(Integer id) {
         if (isEmpty(id)) {
-            throw new ValidationException("Product ID must be informed");
+            throw new ValidationException("The supplier ID must be informed.");
         }
     }
-
 
     public void updateProductStock(ProductStockDTO product) {
         try {
             validateStockUpdateData(product);
             updateStock(product);
-        } catch (Exception e) {
-            log.error("An error occurred while updating product stock: {}", e.getMessage());
-            salesConfirmationSender.sendSalesConfirmationMessage(new SalesConfirmationDTO(product.getSalesId(), SalesStatusEnum.REJECTED));
+        } catch (Exception ex) {
+            log.error("Error while trying to update stock for message with error: {}", ex.getMessage(), ex);
+            var rejectedMessage = new SalesConfirmationDTO(product.getSalesId(), SalesStatusEnum.REJECTED, product.getTransactionid());
+            salesConfirmationSender.sendSalesConfirmationMessage(rejectedMessage);
         }
+    }
+
+    private void validateStockUpdateData(ProductStockDTO product) {
+        if (isEmpty(product)
+                || isEmpty(product.getSalesId())) {
+            throw new ValidationException("The product data and the sales ID must be informed.");
+        }
+        if (isEmpty(product.getProducts())) {
+            throw new ValidationException("The sales' products must be informed.");
+        }
+        product
+                .getProducts()
+                .forEach(salesProduct -> {
+                    if (isEmpty(salesProduct.getQuantity())
+                            || isEmpty(salesProduct.getProductId())) {
+                        throw new ValidationException("The productID and the quantity must be informed.");
+                    }
+                });
     }
 
     @Transactional
@@ -147,84 +210,80 @@ public class ProductService {
                 .getProducts()
                 .forEach(salesProduct -> {
                     var existingProduct = findById(salesProduct.getProductId());
-                    validadeQuantityProduct(salesProduct, existingProduct);
+                    validateQuantityInStock(salesProduct, existingProduct);
                     existingProduct.updateStock(salesProduct.getQuantity());
                     productsForUpdate.add(existingProduct);
                 });
         if (!isEmpty(productsForUpdate)) {
             productRepository.saveAll(productsForUpdate);
-            var approvedMessage = new SalesConfirmationDTO(product.getSalesId(), SalesStatusEnum.APROVED);
+            var approvedMessage = new SalesConfirmationDTO(product.getSalesId(), SalesStatusEnum.APROVED, product.getTransactionid());
             salesConfirmationSender.sendSalesConfirmationMessage(approvedMessage);
         }
     }
 
-    private static void validadeQuantityProduct(ProductQuantityDTO salesProduct, Product existingProduct) {
+    private void validateQuantityInStock(ProductQuantityDTO salesProduct,
+                                         Product existingProduct) {
         if (salesProduct.getQuantity() > existingProduct.getQuantityAvailable()) {
             throw new ValidationException(
-                    String.format("The product %s is out of stock", existingProduct.getId())
-            );
+                    String.format("The product %s is out of stock.", existingProduct.getId()));
         }
     }
 
-    private void validateStockUpdateData(ProductStockDTO product) {
-        if (isEmpty(product) || isEmpty( product.getSalesId())) {
-            throw new ValidationException("The product data or sales id must be provided");
-        }
-        if (isEmpty(product.getProducts())) {
-            throw new ValidationException("Product must be informed");
-        }
-        product.getProducts().forEach(salesProduct -> {
-            if (isEmpty(salesProduct.getQuantity())
-                || isEmpty(salesProduct.getProductId())) {
-                throw new ValidationException("The productID and the quantity must be informed.");
-            }
-        });
-    }
-
-
-    private void validadeProductObject(ProductRequest request) {
-        if (isEmpty(request.getName())) {
-            throw new ValidationException("Product Name must be informed");
-        }
-        if (isEmpty(request.getQuantity())) {
-            throw new ValidationException("Product Quantity must be informed");
-        }
-        if (isEmpty(request.getCategory()) || request.getCategory() == null) {
-            throw new ValidationException("Product Category ID must be informed");
-        }
-        if (isEmpty(request.getSupplier()) || request.getSupplier() == null) {
-            throw new ValidationException("Product Supplier ID must be informed");
-        }
-    }
-
-    public ProductSalesResponse findProductSalesResponse(Integer id) {
+    public ProductSalesResponse findProductSales(Integer id) {
         var product = findById(id);
+        var sales = getSalesByProductId(product.getId());
+        return ProductSalesResponse.of(product, sales.getSalesId());
+    }
+
+    private SalesProductResponse getSalesByProductId(Integer productId) {
         try {
-            var sales = salesClient.findSalesByProductId(product.getId())
-                    .orElseThrow(() -> new ValidationException("The sales was not foundby this Product"));
-            return ProductSalesResponse.of(product, sales.getSalesId());
-        } catch (Exception e) {
-            throw new ValidationException("There was an trying to get the product sales");
+            var currentRequest = getCurrentRequest();
+            var token = currentRequest.getHeader(AUTHORIZATION);
+            var transactionid = currentRequest.getHeader(TRANSACTION_ID);
+            var serviceid = currentRequest.getAttribute(SERVICE_ID);
+            log.info("Sending GET request to orders by productId with data {} | [transactionID: {} | serviceID: {}]",
+                    productId, transactionid, serviceid);
+            var response = salesClient
+                    .findSalesByProductId(productId, token, transactionid)
+                    .orElseThrow(() -> new ValidationException("The sales was not found by this product."));
+            log.info("Recieving response from orders by productId with data {} | [transactionID: {} | serviceID: {}]",
+                    objectMapper.writeValueAsString(response), transactionid, serviceid);
+            return response;
+        } catch (Exception ex) {
+            log.error("Error trying to call Sales-API: {}", ex.getMessage());
+            throw new ValidationException("The sales could not be found.");
         }
     }
 
-    public SuccesResoponse checkProductStock(ProductCheckStockRequest request) {
-        if (isEmpty(request)) {
-            throw new ValidationException("Product ID must be informed");
+    public SuccesResoponse checkProductsStock(ProductCheckStockRequest request) {
+        try {
+            var currentRequest = getCurrentRequest();
+            var transactionid = currentRequest.getHeader(TRANSACTION_ID);
+            var serviceid = currentRequest.getAttribute(SERVICE_ID);
+            log.info("Request to POST product stock with data {} | [transactionID: {} | serviceID: {}]",
+                    objectMapper.writeValueAsString(request), transactionid, serviceid);
+            if (isEmpty(request) || isEmpty(request.getProducts())) {
+                throw new ValidationException("The request data and products must be informed.");
+            }
+            request
+                    .getProducts()
+                    .forEach(this::validateStock);
+            var response = SuccesResoponse.create("The stock is ok!");
+            log.info("Response to POST product stock with data {} | [transactionID: {} | serviceID: {}]",
+                    objectMapper.writeValueAsString(response), transactionid, serviceid);
+            return response;
+        } catch (Exception ex) {
+            throw new ValidationException(ex.getMessage());
         }
-        request.getProducts().forEach(this::validateProductCheckStockRequest);
-
-        return SuccesResoponse.create("The stock is ok!");
     }
 
-    private void validateProductCheckStockRequest(ProductQuantityDTO productQuantityDTO) {
-        if (isEmpty(productQuantityDTO.getProductId()) || isEmpty(productQuantityDTO.getQuantity())) {
-            throw new ValidationException("Product ID and products must be informed");
+    private void validateStock(ProductQuantityDTO productQuantity) {
+        if (isEmpty(productQuantity.getProductId()) || isEmpty(productQuantity.getQuantity())) {
+            throw new ValidationException("Product ID and quantity must be informed.");
         }
-        var product = findById(productQuantityDTO.getProductId());
-        if (productQuantityDTO.getQuantity() > product.getQuantityAvailable()) {
-            throw new ValidationException(String.format("The product %s is out of stock", product.getId()));
+        var product = findById(productQuantity.getProductId());
+        if (productQuantity.getQuantity() > product.getQuantityAvailable()) {
+            throw new ValidationException(String.format("The product %s is out of stock.", product.getId()));
         }
     }
-    
 }
